Move IP header logic to separate struct
